feat: 메인 적용 로딩스피너 전체로 변경
